Enable versioned adaptors - enable dynamic loading of a target Fabric adaptor for gateway or non-gateway modes - add getType() to the blockchain interface
