+++ packages/caliper-fabric/package.json
@@ -23,7 +23,8 @@
     },
     "dependencies": {
         "@hapi/joi": "^15.1.1",
-        "@hyperledger/caliper-core": "0.3.0-unstable"
+        "@hyperledger/caliper-core": "0.3.0-unstable",
+        "semver":"7.1.1"
     },
     "devDependencies": {
         "fabric-ca-client": "^1.4.0",
